Coin Top CC
This is a carbon credit for agriculture based out of the ton of waste. It's specializes in hemp because of the amount of carbon that is sequestered during the time of its growth cycle which is 120 days. This can be used for corn for any other type of excess vegetation that isn't food grade.

Signed-off-by: monkeyman89898989 <buildingguy81@gmail.com>