feat: Download platform tools in image (#202)
Every single verifiable build the process has to download platform
tools. This isn't coming on stdout and it is slow on dodgy wifi, it can
even timeout.
this introduces another pointless critical download opening to supply
chain attack.

Should this be backported to make new images since this does not change
the output?